+++ lib/services/agent_tools/get_app_context_tool.dart
@@ -40,8 +40,8 @@ class GetTagsTool extends AgentTool {
   Future<ToolResult> execute(ToolCall toolCall) async {
     try {
       final offset =
-          toolCall.getInt('offset', defaultValue: 0).clamp(0, 1000000);
-      final limit = toolCall.getInt('limit', defaultValue: 20).clamp(1, 50);
+          toolCall.getInt('offset', defaultValue: 0).clamp(0, 1000000).toInt();
+      final limit = toolCall.getInt('limit', defaultValue: 20).clamp(1, 50).toInt();
 
       final categories = await _databaseService.getCategories();
       final visibleCategories =
